Move `runtime::api` module to Rust runtime
