+++ examples/cryptocurrency/src/lib.rs
@@ -193,8 +193,7 @@ pub mod contracts {
     use exonum::{
         crypto::Hash,
         runtime::{
-            api::ServiceApiBuilder,
-            rust::{CallContext, Service},
+            rust::{api::ServiceApiBuilder, CallContext, Service},
             BlockchainData,
         },
     };
@@ -284,7 +283,7 @@ pub mod contracts {
 pub mod api {
     use exonum::{
         crypto::PublicKey,
-        runtime::api::{self, ServiceApiBuilder, ServiceApiState},
+        runtime::rust::api::{self, ServiceApiBuilder, ServiceApiState},
     };
 
     use crate::schema::{CurrencySchema, Wallet};
